like function hooked into PostCard component
So I can make likes by clicking the heart icon. What I want to do now is style the heart, and write some code so that like posts have a red heart, and if you click it again, the heart will be unliked. That will be the last step I think