Can post comments in the modal now
but cannot make the page rerender and show the new comments. I have a lot of code in the onCommentSubmit function that I'm pretty sure doesn't need to be there. I'm just including all the different functions that might force the page to refresh. I'll have to find a different way.